+++ module-code/app/securesocial/core/IdentityProvider.scala
@@ -129,7 +129,7 @@ abstract class IdentityProvider(application: Application) extends Plugin with Re
   def fillProfile(user: SocialUser):SocialUser
 
   protected def throwMissingPropertiesException() {
-    val msg = "Missing properties for provider '%s'. Verify your configuration file is properly set.".format(id)
+    val msg = "[securesocial] Missing properties for provider '%s'. Verify your configuration file is properly set.".format(id)
     Logger.error(msg)
     throw new RuntimeException(msg)
   }
@@ -140,6 +140,14 @@ object IdentityProvider {
 
   val sslEnabled: Boolean = {
     import Play.current
-    current.configuration.getBoolean("securesocial.ssl").getOrElse(false)
+    val result = current.configuration.getBoolean("securesocial.ssl").getOrElse(false)
+    if ( !result && Play.isProd ) {
+      Logger.warn(
+        "[securesocial] IMPORTANT: Play is running in production mode but you did not turn SSL on for SecureSocial." +
+          "Not using SSL can make it really easy for an attacker to steal your users credentials and/or the " +
+          "authenticator cookie and gain access to the system."
+      )
+    }
+    result
   }
 }