+++ module-code/app/securesocial/core/SecureSocial.scala
@@ -19,11 +19,9 @@ package securesocial.core
 import play.api.mvc._
 import providers.utils.RoutesHelper
 import play.api.i18n.Messages
-import play.api.{Play, Logger}
-import Play.current
+import play.api.Logger
 import play.api.libs.json.Json
 import scala.Some
-import org.joda.time.DateTime
 import play.api.libs.oauth.ServiceInfo
 
 
@@ -42,50 +40,18 @@ case class SecuredRequest[A](user: Identity, request: Request[A]) extends Wrappe
  *    }
  */
 trait SecureSocial extends Controller {
-  val DefaultTimeOutMinutes = 30
-
-  val sessionTimeOutMinutes = Play.current.configuration.getInt("securesocial.sessionTimeOut").getOrElse(
-    DefaultTimeOutMinutes
-  )
-
   /**
    * A Forbidden response for ajax clients
    * @param request
    * @tparam A
    * @return
    */
-  private def ajaxCallNotAuthenticated[A](implicit request: Request[A]): Result = {
-    Unauthorized(Json.toJson(Map("error"->"Credentials required"))).withSession {
-      session - SecureSocial.UserKey - SecureSocial.ProviderKey
-    }.as(JSON)
+  private def ajaxCallNotAuthenticated[A](implicit request: Request[A]): PlainResult = {
+    Unauthorized(Json.toJson(Map("error"->"Credentials required"))).as(JSON)
   }
 
-  private def ajaxCallNotAuthorized[A](implicit request: Request[A]): Result = {
+  private def ajaxCallNotAuthorized[A](implicit request: Request[A]): PlainResult = {
     Forbidden( Json.toJson(Map("error" -> "Not authorized"))).as(JSON)
-  }
-
-  private def touchSession(requestSession: Session, result: PlainResult): Result = {
-    // I can't change the session of a result directly, so I'm getting the cookie
-    // and decoding it from there.
-    // If there is no session in the result, then it's safe to use the
-    // values from the request.
-    val s = for (
-      setCookie <- result.header.headers.get(SET_COOKIE) ;
-      cookie <- Cookies.decode(setCookie).find(_.name  == Session.COOKIE_NAME )
-    ) yield {
-      Session.decodeFromCookie(Some(cookie))
-    }
-    result.withSession(s.getOrElse(requestSession) + SecureSocial.lastAccess)
-  }
-
-  def lastAccessFromSession(session: Session): Option[DateTime] = {
-    session.data.get(SecureSocial.LastAccessKey).map {
-      DateTime.parse(_)
-    }
-  }
-
-  def isSessionExpired(lastAccess: DateTime): Boolean = {
-    DateTime.now().isAfter( lastAccess.plusMinutes(sessionTimeOutMinutes))
   }
 
   /**
@@ -105,23 +71,17 @@ trait SecureSocial extends Controller {
     implicit request => {
 
       val result = for (
-        lastAccess <- lastAccessFromSession(session) ;
-        userId <- SecureSocial.userFromSession if !isSessionExpired(lastAccess) ;
-        user <- UserService.find(userId)
+        authenticator <- SecureSocial.authenticatorFromRequest ;
+        user <- UserService.find(authenticator.userId)
       ) yield {
+        touch(authenticator)
         if ( authorize.isEmpty || authorize.get.isAuthorized(user)) {
-          f(SecuredRequest(user, request)) match {
-            case plainResult: PlainResult => {
-              touchSession(request.session, plainResult)
-            }
-            case r => r
-          }
-
+          f(SecuredRequest(user, request))
         } else {
           if ( ajaxCall ) {
             ajaxCallNotAuthorized(request)
           } else {
-            Redirect(RoutesHelper.notAuthorized.absoluteURL(IdentityProvider.sslEnabled), UNAUTHORIZED)
+            Redirect(RoutesHelper.notAuthorized.absoluteURL(IdentityProvider.sslEnabled))
           }
         }
       }
@@ -130,16 +90,14 @@ trait SecureSocial extends Controller {
         if ( Logger.isDebugEnabled ) {
           Logger.debug("[securesocial] anonymous user trying to access : '%s'".format(request.uri))
         }
-        if ( ajaxCall ) {
+        val response = if ( ajaxCall ) {
           ajaxCallNotAuthenticated(request)
         } else {
           Redirect(RoutesHelper.login()).flashing("error" -> Messages("securesocial.loginRequired")).withSession(
             session + (SecureSocial.OriginalUrlKey -> request.uri)
-              - SecureSocial.UserKey
-              - SecureSocial.ProviderKey
-              - SecureSocial.LastAccessKey
           )
         }
+        response.discardingCookies(Authenticator.cookieName)
       })
     }
   }
@@ -206,13 +164,14 @@ trait SecureSocial extends Controller {
    */
   def UserAwareAction[A](p: BodyParser[A])(f: RequestWithUser[A] => Result) = Action(p) {
     implicit request => {
-      val user = SecureSocial.currentUser
-      f(RequestWithUser(user, request)) match {
-        case plainResult: PlainResult if user.isDefined => {
-          touchSession(request.session, plainResult)
-        }
-        case r => r
+      val user = for (
+        authenticator <- SecureSocial.authenticatorFromRequest ;
+        user <- UserService.find(authenticator.userId)
+      ) yield {
+        touch(authenticator)
+        user
       }
+      f(RequestWithUser(user, request))
     }
   }
 
@@ -224,31 +183,34 @@ trait SecureSocial extends Controller {
   def UserAwareAction(f: RequestWithUser[AnyContent] => Result): Action[AnyContent] = {
     UserAwareAction(parse.anyContent)(f)
   }
+
+  def touch(authenticator: Authenticator) {
+    Authenticator.save(authenticator.touch)
+  }
 }
 
 object SecureSocial {
-  val UserKey = "securesocial.user"
-  val ProviderKey = "securesocial.provider"
   val OriginalUrlKey = "securesocial.originalUrl"
-  val LastAccessKey = "securesocial.lastAccess"
-  val SessionTimeOut = "securesocial.sessionTimeOutMinutes"
 
-  def lastAccess = (SecureSocial.LastAccessKey -> DateTime.now().toString())
+  def authenticatorFromRequest(implicit request: RequestHeader): Option[Authenticator] = {
+    val result = for {
+      cookie <- request.cookies.get(Authenticator.cookieName) ;
+      maybeAuthenticator <- Authenticator.find(cookie.value).fold(e => None, Some(_)) ;
+      authenticator <- maybeAuthenticator
+    } yield {
+      authenticator
+    }
 
-
-  /**
-   * Build a UserId object from the session data
-   *
-   * @param request
-   * @tparam A
-   * @return
-   */
-  def userFromSession[A](implicit request: RequestHeader):Option[UserId] = {
-    for (
-      userId <- request.session.get(SecureSocial.UserKey);
-      providerId <- request.session.get(SecureSocial.ProviderKey)
-    ) yield {
-      UserId(userId, providerId)
+    result match {
+      case Some(a) => {
+        if ( !a.isValid ) {
+          Authenticator.delete(a.id)
+          None
+        } else {
+          Some(a)
+        }
+      }
+      case None => None
     }
   }
 
@@ -262,8 +224,8 @@ object SecureSocial {
    */
   def currentUser[A](implicit request: RequestHeader):Option[Identity] = {
     for (
-      userId <- userFromSession ;
-      user <- UserService.find(userId)
+      authenticator <- authenticatorFromRequest ;
+      user <- UserService.find(authenticator.userId)
     ) yield {
       user
     }